+++ Practica1/HolaSDL/Game.h
@@ -45,6 +45,6 @@ public:
 	void run();
 	void render();
 	void update();
-	bool collides(Vector2D& pos, int size, int& angle);
+	bool collides(Vector2D& pos, int size, double& angle);
 	void handleEvents();
 };
